DALLE_mini/requirements.txt
min-dalle>=1.0.0
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.0.0
matplotlib>=3.5.0
transformers>=4.20.0
